Use correct default remote for fetching in submodule.
`git submodule update` first tries a regular `git fetch`
to fetch the commit, but when that doesn't retrieve the commit
it wants, it tries `git fetch <remote> <commit>`.
For `<remote>`, it used the wrong default remote: the default
remote of the outer repository, rather than the default remote
of the submodule.

Signed-off-by: Mara Bos <mara@marabos.nl>